// lib/tax-forms/frontend-types.ts
export enum TaxFormType {
  W9 = 'W9',
  W8_BEN = 'W8_BEN',
  W8_BEN_E = 'W8_BEN_E',
}

export enum SubmitterType {
  Individual = 'Individual',
  Business = 'Business',
}
// ...
export type W8BenTaxFormValues = {
  isUSPersonOrEntity?: boolean;
  submitterType?: SubmitterType;
  formType?: TaxFormType;
  email?: string;
  signer?: { firstName?: string; middleName?: string; lastName?: string };
  isSigned?: boolean;
  beneficialOwner?: { firstName?: string; middleName?: string; lastName?: string };
  taxpayerIdentificationNumberTypeUS?: 'SSN' | 'ITIN';
  taxpayerIdentificationNumberUS?: string;
  taxpayerIdentificationNumberForeign?: string;
  dateOfBirth?: string;
  claimsSpecialRatesAndConditions?: boolean;
  claimsArticleAndParagraph?: string;
  claimsRate?: number;
  claimsIncomeType?: string;
  claimsExplanation?: string;
  hasTaxTreatySpecialRatesAndConditions?: boolean;
  certifiesResidentCountry?: boolean;
  hasConfirmedTOS?: boolean;
  countryOfCitizenship?: string;
  residenceAddress?: {
    country?: string;
    structured?: { address1?: string; address2?: string; city?: string; zone?: string; postalCode?: string };
  };
  mailingAddress?: {
    country?: string;
    structured?: { address1?: string; address2?: string; city?: string; zone?: string; postalCode?: string };
  };
  isSignerTheBeneficialOwner?: boolean;
  signerCapacity?: string;
};